day five

+++ advent_of_code_2022_cpp/day_four/day_four.cpp
@@ -62,5 +62,5 @@ int calculate_in_range_pairs(vector<string> inputPairs)
         }
     }
     
-    return inputPairs.size() - result;
+    return (int)(inputPairs.size() - result);
 }
